Add token-provider class
A token allows a query to identify a particular simulation.

token-provider returns a unique token-id, given two parameters: a
multiplier (which denotes at what speed the matches will be simulated)
and a timestamp (which denotes when the simulation started).

The token will time out after 105 minutes of simulated time (which
may be less that 105 minutes of real time if the multiplier is
larger than 1).